% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow_separation.R
\name{digital_filter}
\alias{digital_filter}
\title{Digital filter}
\usage{
digital_filter(y, alpha)
}
\arguments{
\item{y}{Baseflow timeseries}

\item{alpha}{Filter parameters (0.9 - 0.95)}
}
\description{
Digital filter for baseflow separation
}
